rename some directory
