+++ api/pkg/s3/datastore/huawei/hws.go
